requirements.txt
Flask
flask-cors
tensorflow
pillow
pandas
numpy
joblib